pyproject.toml
[tool.poetry]
name = "stealer-parser"
version = "1.0.0"
description = "Infostealer logs parser."
authors = ["matya <135035586+m4tya@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
coloredlogs = "^15.0.1"
py7zr = "^0.21.0"
rarfile = "^4.2"
verboselogs = "^1.7"
psycopg2-binary = "^2.9.9"
dependency-injector = "^4.48.1"
pydantic-settings = "^2.10.1"
PyYAML = "^6.0.2"

[tool.poetry.scripts]
stealer_parser = "stealer_parser.main:main"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pycodestyle = "^2.12.0"
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
